+++ src/main/java/com/invadermonky/justenoughmagiculture/JustEnoughMagiculture.java
@@ -3,6 +3,7 @@ package com.invadermonky.justenoughmagiculture;
 import com.invadermonky.justenoughmagiculture.proxy.CommonProxy;
 import com.invadermonky.justenoughmagiculture.util.LogHelper;
 import com.invadermonky.justenoughmagiculture.util.ModIds.ConstantIds;
+import com.invadermonky.justenoughmagiculture.util.ModIds.ConstantVersions;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.SidedProxy;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
@@ -21,7 +22,7 @@ public class JustEnoughMagiculture {
     public static final String MOD_ID = "justenoughmagiculture";
     public static final String MOD_ALIAS = "jem";
     public static final String MOD_NAME = "Just Enough Magiculture";
-    public static final String MOD_VERSION = "1.12.2-1.1.3";
+    public static final String MOD_VERSION = "1.12.2-1.1.4";
     public static final String MC_VERSION = "[1.12.2]";
     public static final String DEPENDENCIES = (
             "required-after:" + ConstantIds.JUSTENOUGHITEMS + "@[4.7.0,)" +
@@ -29,7 +30,7 @@ public class JustEnoughMagiculture {
                     ";after:" + ConstantIds.ANIMANIA +
                     ";after:" + ConstantIds.ATUM +
                     ";after:" + ConstantIds.BEAR_WITH_ME +
-                    ";after:" + ConstantIds.BEAST_SLAYER +
+                    ";after:" + ConstantIds.BEAST_SLAYER_BETA + "@" + ConstantVersions.BEAST_SLAYER_BETA +
                     ";after:" + ConstantIds.BETWEENLANDS +
                     ";after:" + ConstantIds.BEWITCHMENT +
                     ";after:" + ConstantIds.BOTANIA +