+++ src/main/java/com/invadermonky/justenoughmagiculture/proxy/ClientProxy.java
@@ -2,6 +2,7 @@ package com.invadermonky.justenoughmagiculture.proxy;
 
 import com.invadermonky.justenoughmagiculture.init.InitIntegration;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 
 public class ClientProxy extends CommonProxy {
@@ -14,6 +15,11 @@ public class ClientProxy extends CommonProxy {
     @Override
     public void init(FMLInitializationEvent event) {
         super.init(event);
+    }
+
+    @Override
+    public void postInit(FMLPostInitializationEvent event) {
+        super.postInit(event);
         InitIntegration.init();
     }
 }